+++ rmit-combined/metadata.rb
@@ -9,6 +9,12 @@ description      'Installs/Configures `rmit-combined` for cwb'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.4'
 
+# CWB
 depends 'cwb', '~> 0.1.0'
+
+# External
 depends 'apt', '~> 6.0.1'
 depends 'build-essential', '~> 8.0.0'
+
+# Internal
+depends 'stop-services', '~> 0.1.0'